% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ontologies.R
\name{graph_ontology}
\alias{graph_ontology}
\title{Graph the MOLON ontology}
\usage{
graph_ontology()
}
\value{
A \code{ggplot2} object representing the ontology graph.
}
\description{
Visualise the MOLON ontology that describes mechanisms
driving distinct mutational profiles in cancer.
}
\examples{
ontology_graph <- graph_ontologies()
print(ontology_graph)
}
