% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ontologies.R
\name{load_ontology}
\alias{load_ontology}
\title{Load Ontologies}
\usage{
load_ontology()
}
\value{
A data frame containing the ontologies.
}
\description{
This function loads the 'molon' ontology which describes mechanisms driving
mutational profiles
}
\examples{
ontologies <- load_ontologies()
head(ontologies)
}
